+++ components/demo-firebase/jest.config.ts
@@ -1,14 +1,7 @@
+(global as any).appTestType = 'firebase';
+
 module.exports = {
   displayName: 'demo-firebase',
-  globals: {
-    'ts-jest': {
-      tsconfig: '<rootDir>/tsconfig.spec.json'
-    }
-  },
-  transform: {
-    '^.+\\.[tj]s$': 'ts-jest'
-  },
-  moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../coverage/apps/demo-firebase',
-  preset: '../../jest.preset.ts'
+  preset: '../../jest.preset.ts',
+  coverageDirectory: '../../coverage/apps/demo-firebase'
 };