+++ jest.setup.firebase.ts
@@ -1,3 +1,5 @@
+import './jest.setup.nestjs';
+
 /**
  * Initialize FIREBASE_CONFIG for tests to prevent a warning for firebase not being initialized.
  *