+++ packages/dbx-analytics/jest.config.ts
@@ -1,17 +1,7 @@
+(global as any).appTestType = 'angular';
+
 module.exports = {
   displayName: 'dbx-analytics',
-  setupFilesAfterEnv: ['<rootDir>/src/test-setup.ts'],
-  globals: {
-    'ts-jest': {
-      tsconfig: '<rootDir>/tsconfig.spec.json',
-      stringifyContentPathRegex: '\\.(html|svg)$'
-    }
-  },
-  coverageDirectory: '../../coverage/packages/dbx-analytics',
-  transform: {
-    '^.+\\.(ts|mjs|js|html)$': 'jest-preset-angular'
-  },
-  transformIgnorePatterns: ['node_modules/(?!.*\\.mjs$)'],
-  snapshotSerializers: ['jest-preset-angular/build/serializers/no-ng-attributes', 'jest-preset-angular/build/serializers/ng-snapshot', 'jest-preset-angular/build/serializers/html-comment'],
-  preset: '../../jest.preset.ts'
+  preset: '../../jest.preset.ts',
+  coverageDirectory: '../../coverage/packages/dbx-analytics'
 };