+++ packages/nestjs/jest.config.ts
@@ -1,15 +1,7 @@
+(global as any).appTestType = 'nestjs';
+
 module.exports = {
   displayName: 'nestjs',
   preset: '../../jest.preset.ts',
-  globals: {
-    'ts-jest': {
-      tsconfig: '<rootDir>/tsconfig.spec.json'
-    }
-  },
-  testEnvironment: 'node',
-  transform: {
-    '^.+\\.[tj]s$': 'ts-jest'
-  },
-  moduleFileExtensions: ['ts', 'js', 'html'],
   coverageDirectory: '../../coverage/packages/nestjs'
 };